Move account delete button to account details
