+++ src/components/AccountDetailsHeader.jsx
@@ -1,9 +1,32 @@
 import React from 'react';
+import { FaTrash } from 'react-icons/fa';
 
 export default function AccountDetailsHeader({ account, transactions }) {
+    function handleSubmit() {
+        deleteAccount();
+    }
+
+    async function deleteAccount() {
+        try {
+            const options = {
+                method: 'DELETE'
+            };
+
+            const response = await fetch(`http://localhost:4000/accounts/${account.account_id}`, options);
+
+        } catch (error) {
+            console.error(error.message);
+        }
+    }
+
     return (
         <div className='header details-header account-details-header'>
             <h2 className="details-title">{`${account.name} Details`}</h2>
+            <form onSubmit={handleSubmit} className='delete-form account-delete-form'>
+                <button type="submit">
+                    <FaTrash />
+                </button>
+            </form>
             <table>
                 <thead>
                     <tr>
